Move interpret_actions back to HCNN agent
select_actions should return both the agent_action for ReplayMemory and actions that environment can use directly to place blocks
